add options to search for subject and remove if entryid is known

+++ delete-items/delete-items.py
@@ -2,21 +2,20 @@
 # -*- coding: utf-8 -*-
 # vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
 #
-try:
-    import kopano
-except ImportError:
-    import zarafa as kopano
-    print 'Script is tested with Kopano, error handling will not work with Zarafa'
+import kopano
 import sys
-from MAPI.Util import *
+from MAPI.Tags import *
 from datetime import datetime
 
 def opt_args():
-    parser = kopano.parser('skpfucm')
+    parser = kopano.parser('skpfucmUP')
 
     parser.add_option("--from", dest="from_date", action="store", help="Remove items older than x (YYYY-MM-DD)")
     parser.add_option("--until", dest="until_date", action="store", help="Remove item till this date (YYYY-MM-DD)")
     parser.add_option("--all", dest="all", action="store_true", help="Remove all items")
+    parser.add_option("--subject", dest="subject", action="store", help="Subject of item")
+    parser.add_option("--entryid", dest="entryid", action="store", help="entryid of item")
+
     parser.add_option("--dry-run", dest="dry_run", action="store_true", help="Dry run")
     return parser.parse_args()
 
@@ -24,12 +23,12 @@ def main():
     options, args = opt_args()
 
     if not options.users :
-        print 'Usage:\n%s -u <username> ' % sys.argv[0]
+        print('Usage:\n{} -u <username>'.format(sys.argv[0]))
         sys.exit(1)
 
-    if not options.from_date and not options.until_date:
-        print 'usage:\n%s -u <username> --from* YYYY-MM-D --until* YYYY-MM-DD \n' \
-              '* only one parameter is required ' % (sys.argv[0])
+    if (not options.from_date and not options.until_date) and (not options.subject and not options.entryid):
+        print('usage:\n{} -u <username> --from* YYYY-MM-D --until* YYYY-MM-DD \n'
+              '* only one parameter is required '.format(sys.argv[0]))
         sys.exit(1)
 
     current_date = datetime.now()
@@ -41,12 +40,20 @@ def main():
         until_date = datetime.strptime(options.until_date, '%Y-%m-%d')
     else:
         until_date = datetime.strptime('2038-01-19', '%Y-%m-%d')
+    server= kopano.Server(options)
 
-    for user in kopano.Server(options).users():
+    if options.entryid:
+        user = server.user(options.users[0])
+        item = user.store.item(options.entryid)
+        print('Deleting item {}'.format(item.subject))
+        user.store.delete(item)
+        sys.exit(0)
+
+    for user in server.users():
         item_delete = 0
-        print 'Runnig for user %s' % user.name
+        print('Runnig for user {}'.format(user.name))
         for folder in user.store.folders():
-            print 'Search items in %s' % folder.name
+            print('Search items in {}'.format(folder.name))
             for item in folder.items():
                 if item.received == None:
                     received_date = item.created
@@ -54,12 +61,15 @@ def main():
                     received_date = item.received
 
                 if received_date >= from_date and received_date <= until_date:
+                    if options.subject  and options.subject != item.subject:
+                        continue
                     item_delete += 1
                     if options.dry_run:
-                        print '{:150} {}'.format(item.subject, item.received)
+                        print('{} {} with entryid {}'.format(item.subject, item.received, item.entryid))
                     else:
+                        print('removing item {}'.format(item.subject))
                         folder.delete(item)
 
-        print '\nDeleted items : %s' % item_delete
+        print('\nDeleted items :{}'.format(item_delete))
 if __name__ == "__main__":
     main()
